rename grade list component
